Removed 'header' 'body' arguments, cleaned up code
Report scripts are no longer given any arguments. Everything is
processed via the standard output of the script. The first line
is header, everything else is the body.

Added setting default checking, only the settings that need
to be changed have to be set. Everything else will be set to program
defaults.

Cleaned up apache and nginx reports to check if the file exists
instead of checking the error status of 'tail'.

Syslog report: reversed lines when tailing syslog or messages to
be consistant with the journalctl command.