Local config takes priority, moved logfile and tempmail to config
IF a configuration file is located in the same directory as the script,
that will take priority over a config in /etc. This allows for the script
to be portable without unexpected results due to an existing config
file.

The variables LOGFILE and TEMPMAIL can now be defined in the config
file with default values being the same as the precious hard-coded
ones.