Safely import config file, no export, smart email disable
The configuration file is no longer sourced in order to avoid
potential security issues (of course if someone could modify the
config, they can probably modify running reports). The config is read,
sanitized, and then values are declared.

Removed the export command from the sample conf file as it's no
longer needed. Actually is hasn't been needed at all.

EMAIL_REPORT will be set to false automatically if EMAIL_TO is empty.
It will be set to true if set in the config file of course it still
wouldn't mail if EMAIL_TO is still empty.